+++ src/onegov.org/onegov/org/directives.py
@@ -19,3 +19,22 @@ class HomepageWidgetAction(Action):
         widget.tag = self.tag  # keep redundantly for ease of access
 
         homepage_widget_registry[self.tag] = widget
+
+
+class ExportAction(Action):
+    """ Register an export. """
+
+    config = {
+        'export_registry': dict
+    }
+
+    def __init__(self, id, **kwargs):
+        self.id = id
+        self.kwargs = kwargs
+        self.kwargs['id'] = id
+
+    def identifier(self, export_registry):
+        return self.id
+
+    def perform(self, cls, export_registry):
+        export_registry[self.id] = cls(**self.kwargs)